refactor(cockpit): extract code health module
